requirements.txt
discord.py==2.4.0
openai==1.40.8
python-dotenv==1.0.0
sqlalchemy
openpyxl
cx_Oracle
scikit-learn